+++ spec/jac/string_eval_spec.rb
@@ -40,6 +40,23 @@ describe Configuration do
       end
     end
 
+    context 'at top level of profile values can be referenced without c' do
+      let(:conf) do
+        <<-CONFIG.strip_indent
+        default:
+          project: sandbox-release
+          project_c_eval: "\#{c.project}"
+          project_self_eval: "\#{project}"
+        CONFIG
+      end
+
+      it do
+        c = Configuration.read('default', conf)
+        # expect(c.project_c_eval).to eq(c.project)
+        expect(c.project_self_eval).to eq(c.project)
+      end
+    end
+
     context 'list elements can be accesed via `self[idx]`' do
       let(:conf) do
         <<-CONFIG.strip_indent